ref: Refactor file handling
The `file` of a returned `SourceLocation` is now no longer just the name
that you can then use to look up the rest of the information, but the whole structure.
This change is mirrored in the raw data structures: the `file_idx` no longer points to the file name,
but the entire file structure. This makes it necessary to restructure `SmCacheWriter::new`
so that file insertion happens first.

I also added some convenience methods and renamed some things.