ref: Switch JS parser from rslint to swc
The rslint parser is unmaintained and has a hard limit on how much code
it can parse.

This changes the internal parser so swc, which is maintained and production ready.

Instead of having untyped SyntaxNodes that allow free navigation within
the hierarchy, swc is rather based around typed AST nodes and a visitor pattern.

Conversion between the two is mostly straight forward but does involve quite some manual work.